23042021 sore

+++ app/Controllers/Setting.php
@@ -10,9 +10,10 @@ class Setting extends ResourceController
 
 	public function index()
 	{
+        $id = '1';
         $data = [
-            'status' => '1',
-            'data' => $this->model->findAll()
+            'status' => '200',
+            'data' => $this->model->getSetting($id)
         ];
 
         return $this->respond($data, 200);
@@ -21,11 +22,11 @@ class Setting extends ResourceController
     public function show($id = null)
     {
         $data = [
-            'status' => '1',
+            'status' => '200',
             'data' => $this->model->find($id)
         ];
 
-        return $this->response->setStatusCode(200)->setJSON($data);
+        return $this->respond($data, 200);
     }
 
     public function create()
@@ -43,13 +44,13 @@ class Setting extends ResourceController
         if ($data > 0) {
             $this->model->save($data);
             $response = [
-                'status' => '1',
+                'status' => '201',
                 'data' => 'Success Post Data'
             ];
             return $this->respond($response, 201);
         } else {
             $response = [
-                'status' => '0',
+                'status' => '422',
                 'data' => 'Failed Post Data'
             ];
             return $this->respond($response, 422);
@@ -74,21 +75,21 @@ class Setting extends ResourceController
                 $this->model->update($id, $data);
 
                 $response = [
-                    'status' => '1',
+                    'status' => '200',
                     'data' => 'Success Update data'
                 ];
                 return $this->respond($response, 200);
             } 
 
             $response = [
-                'status' => '0',
+                'status' => '422',
                 'data' => 'Failed Update Data'
             ];
             return $this->respond($response, 422);
         }
 
         $response = [
-            'status' => '0',
+            'status' => '404',
             'data' => 'Failed Update Data'
         ];
         return $this->respond($response, 404);
@@ -96,40 +97,52 @@ class Setting extends ResourceController
 
     public function delete($id = null)
     {
-        if ($this->model->find($id)) {
+        if ($this->request)
+        {
+            //get request from Reactjs
+            if($this->request->getJSON()) {
+                if ($this->model->find($id)) {
 
-            $delete = $this->model->delete($id);
-    
-            if ($delete) {
-                $response = [
-                    'status' => '1',
-                    'data' => 'Success Delete data'
-                ];
-                return $this->respond($response, 200);
-            } 
-
-            try {
-                $this->model->delete($id);
-                $response = [
-                    'status' => '1',
-                    'data' => 'Success Delete data'
-                ];
-                return $this->respond($response, 200);
-            } catch (\Exception $e) {
-                $response = [
-                    'status' => '0',
-                    'data' => 'Failed Delete Data'
-                ];
-                return $this->respond($response, 422);
-            }
+                    $delete = $this->model->delete($id);
             
-        }
+                    if ($delete) {
+                        $response = [
+                            'status' => '200',
+                            'data' => 'Success Delete Data'
+                        ];
+                        return $this->respond($response, 200);
+                    } 
 
-        $response = [
-            'status' => '0',
-            'data' => 'Failed Delete Data'
-        ];
-        return $this->respond($response, 404);
+                    //try {
+                        //$this->model->delete($id);
+                        //$response = [
+                            //'status' => '1',
+                            //'data' => 'Success Delete data'
+                        //];
+                        //return $this->respond($response, 200);
+                    //} catch (\Exception $e) {
+                        //$response = [
+                            //'status' => '0',
+                            //'data' => 'Failed Delete Data'
+                        //];
+                        //return $this->respond($response, 422);
+                    //}
+                    
+                }
+            } else {
+                $response = [
+                    'status' => '405',
+                    'data' => 'Data Not Found'
+                ];
+                return $this->respond($response, 405);
+            }
+        } else {
+            $response = [
+                'status' => '404',
+                'data' => 'Data Not Found'
+            ];
+            return $this->respond($response, 404);
+        }
     }
     
 }
